requirements.txt
Pillow
gensim
numpy
tensorflow